Emil-7/EmilO-task7-task-202410141648/.hidden_tasks/UserFileReader.java
import java.util.Map;
import java.util.HashMap;

public class UserFileReader {

    public Map<String, String> readUsers(String filename) {
        Map<String, String> userMap = new HashMap<>();

        try (BufferedReader reader = new BufferedReader(new FileReader(filename))) {
            String line;
            while ((line = reader.readLine()) != null) {
                String[] parts = line.split(",");
                if (parts.length == 2) {
                    String username = parts[0].trim();
                    String password = parts[1].trim();
                    userMap.put(username, password);
                } else {
                    System.out.println("Skipping invalid line: " + line);
                }
            }
        } catch (IOException e) {
            System.err.println("Error reading file: " + e.getMessage());
        }

        return userMap;
    }
}